Add files via upload
The file form.php contains the form that is being included on the index.php.
The file thanks.php is the page that will be displayed after the form submit from the user, containing a customized message, in accordance to the data got by POST method. 
The file style.css will apply the style in all pages.